requirements.txt
aiohttp==3.6.2
alpha-vantage==2.2.0
altgraph==0.17
APScheduler==3.6.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
blis==0.7.4
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
cymem==2.0.5
dateutils==0.6.11
Django==3.1.2
django-crispy-forms==1.9.2
django-crontab==0.7.1
django-widget-tweaks==1.4.8
gunicorn==20.0.4
idna==2.10
iniconfig==1.0.1
kiwisolver==1.2.0
Library==0.0.0
matplotlib==3.3.2
more-itertools==8.5.0
multidict==4.7.6
murmurhash==1.0.5
newsapi-python==0.2.6
nsepython==0.0.94
nsetools==1.0.11
numpy==1.19.2
packaging==20.4
pandas==1.1.2
Pillow==7.2.0
Pint==0.16.1
plac==0.9.6
pluggy==0.13.1
preshed==3.0.5
py==1.9.0
pyalgotrading==2020.8.7
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pyparsing==2.4.7
pytest==6.0.2
pytest-asyncio==0.10.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.1.2
requests==2.24.0
schedule==0.6.0
six==1.15.0
sqlparse==0.4.1
srsly==1.0.5
thinc==7.4.5
toml==0.10.1
tqdm==4.55.0
tzlocal==2.1
urllib3==1.25.10
vim==0.0.1
wasabi==0.8.0
whitenoise==5.2.0
yarl==1.5.1
